refactor(test): address Bot Review round 3 on #293 — shared controller harness
- Promote the agent-process-controller test harness to
  tests/unit/tui/support/ now that two notice tests genuinely share the
  wiring contract (per the style guide's "promote at the second consumer"
  rule); both instruction-warning and roots-warning suites import it.
- Align ordering across the two create-session paths: session store first
  (it owns .vesicle), then ensureProjectRoots, in both
  initializeSessionIdentity and bootstrapTurn.
- Type-narrow the roots-event assertion with an Extract predicate instead
  of a chained && inside expect().

Refs #291
