+++ tests/unit/tui/tui-instruction-warning.test.ts
@@ -1,9 +1,7 @@
-import { createRoot, createSignal } from "solid-js";
+import { createRoot } from "solid-js";
 import { describe, expect, test } from "bun:test";
-import { createAgentProcessController } from "../../../src/tui/agent-process-controller";
 import type { AgentLoopEvent } from "../../../src/core/agent-loop/run";
-import type { ActivityEntry, AgentCardState, Message } from "../../../src/tui/types";
-import type { BackgroundProcessState } from "../../../src/core/process/manager";
+import { agentProcessControllerHarness } from "./support/agent-process-controller-harness";
 
 type InstructionWarningEvent = Extract<AgentLoopEvent, { type: "instruction_warning" }>;
 type InstructionDiagnostic = InstructionWarningEvent["diagnostics"][number];
@@ -19,37 +17,7 @@ function warningEvent(
   return { type: "instruction_warning", sessionId, engine: "runtime", diagnostics };
 }
 
-function harness(sessionId = "session-1") {
-  const [messages, setMessages] = createSignal<Message[]>([]);
-  const [activity, setActivity] = createSignal<ActivityEntry[]>([]);
-  const [background, setBackground] = createSignal<BackgroundProcessState[]>([]);
-  const [cards, setCards] = createSignal<AgentCardState[]>([]);
-  const [status, setStatus] = createSignal("");
-  const [streamingAssistant, setStreamingAssistant] = createSignal("");
-  const [streamingReasoning, setStreamingReasoning] = createSignal("");
-  const [lastToolContent, setLastToolContent] = createSignal<string | null>(null);
-  const controller = createAgentProcessController({
-    sessionId: () => sessionId,
-    busy: () => false,
-    activeEngine: () => "runtime",
-    activeModel: () => "test-model",
-    backgroundProcesses: background,
-    setBackgroundProcesses: setBackground,
-    setAgentCards: setCards,
-    setMessages,
-    setActivity,
-    setStatus,
-    setStreamingAssistant,
-    setStreamingReasoning,
-    setLastDisplayedToolAssistantContent: setLastToolContent,
-    markTurnSawResponse: () => undefined,
-    recordResponseUsage: () => undefined,
-    recordIndependentAgentUsage: () => undefined,
-    assetDriftKey: () => undefined,
-    setAssetDriftKey: () => undefined,
-  });
-  return { controller, messages, activity, status, background, cards, streamingAssistant, streamingReasoning, lastToolContent };
-}
+const harness = agentProcessControllerHarness;
 
 describe("instruction warning notice", () => {
   test("dedups identical warnings across events and re-notifies when the diagnostic changes", () => createRoot((dispose) => {
